Virtual Piano
Virtual Piano is a graduate project for track Frontend Core.

You will learn how to write HTML documents and work with the most popular CSS properties. You will also learn how to use JavaScript to make event handlers for browser events and create and insert new elements into documents.

One of the many things that a front-end developer has to take care of is binding the functionality of a web application to the keys on the keyboard. In order to develop and strengthen the important skills of working with keyboard keys, we will make a virtual piano in this project.