Stage 1/5: Add markup
To go through the first step, change the HTML page title and then assign the keys for your virtual piano.

The <kbd> tag is used to denote the keyboard input or a key name (you can read more about it on MDN). Use this paired tag in the body of the document to specify seven keys on the keyboard: A, S, D, F, G, H, and J. These will be the white keys on the piano.

If you open this HTML file in your browser, you'll see the names of the keyboard keys. These keys do not make sounds yet, but you're one step closer to creating a working virtual piano.

Objectives

    Change the page title to "Virtual Piano".
    Add 7 <kbd> tags to the document body and enclose letters A, S, D, F, G, H, and J in them.